Remove op:bootstrap and instead always export at `0`
This removes the `op:bootstrap` operation which used to ask the remote
node make available the bootstrap object and would setup a promise. That
is almost always unnessesery as the bootstrap object is needed to do
anything useful with CapTP and so must always be issued.

This instead removes the operation and just makes it required that after
a successful `op:start-session` exchange the bootstrap object be
exported at position `0` which simplifies the spec.